+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Added
 
+- BC-121 add bucket and storage provider to file documents
 - BC-164 - refactored the supportJWTservice and added roles information to its jwt data
 - BC-279 - introduce console application interface to call providers via console command
 - BC-9 add entity and repo for course dashboards